Switch emoji line to oneline.RenderEmoji (#345, #1222, #1225)
